Move managers to models package
