+++ sponsors/models/managers.py
